% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gain_and_offset.R
\name{calculate_gain_and_offset}
\alias{calculate_gain_and_offset}
\title{Title}
\usage{
calculate_gain_and_offset(std1, std2)
}
\arguments{
\item{std1}{}

\item{std2}{}
}
\value{

}
\description{
Title
}
