% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_ambient_carbon_linreg.R
\name{calibrate_ambient_carbon_linreg}
\alias{calibrate_ambient_carbon_linreg}
\title{calibrate_ambient_carbon_isotopes}
\usage{
calibrate_ambient_carbon_linreg(amb.data.list, caldf, outname, site, file,
  force.to.end = TRUE, force.to.beginning = TRUE)
}
\arguments{
\item{amb.data.list}{}

\item{caldf}{}

\item{outname}{}

\item{site}{}

\item{file}{}

\item{force.to.end}{}

\item{force.to.beginning}{}

\item{method}{}
}
\description{
calibrate_ambient_carbon_isotopes
}
\examples{


}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
