% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_carbon_bymonth.R
\name{calibrate_carbon_bymonth}
\alias{calibrate_carbon_bymonth}
\title{calibrate_carbon_bymonth}
\usage{
calibrate_carbon_bymonth(
  inname,
  outname,
  site,
  method = "Bowling_2003",
  calibration_half_width = 0.5,
  force_cal_to_beginning = TRUE,
  force_cal_to_end = TRUE,
  gap_fill_parameters = FALSE,
  filter_ambient = TRUE,
  r2_thres = 0.95,
  correct_refData = TRUE,
  write_to_file = TRUE
)
}
\arguments{
\item{inname}{Name of the input file. (character)}

\item{outname}{Name of the output file. (character)}

\item{site}{Four letter NEON site code for site being processed. (character)}

\item{method}{Are we using the Bowling et al. 2003 method
("Bowling_2003") or direct linear regression of
d13C and CO2 mole fractions ("linreg")?}

\item{calibration_half_width}{Determines the period (in days)
from which reference data are selected (period
is 2*calibration_half_width).}

\item{force_cal_to_beginning}{Extend first calibration to the beginning
of the file? (CURRENTLY NOT USED)}

\item{force_cal_to_end}{Extend last calibration to the end of the file?
(CURRENTLY NOT USED)}

\item{gap_fill_parameters}{Should function attempt to 'gap-fill' across a 
bad calibration by carrying the last known good calibration forward?
Implementation is fairly primitive currently, as it only carries 
the last known good calibration that's available forward rather
than interpolating, etc. Default FALSE.}

\item{filter_ambient}{Apply the median absolute deviation filter (Brock 86)
to remove impulse spikes in output ambient data?
(logical; default true)}

\item{r2_thres}{Minimum r2 threshold of an "acceptable" calibration. Acts to
remove calibration periods where a measurement error makes
relationship nonlinear. Default = 0.95}

\item{correct_refData}{NEON has indicated there are a few instances where
reported d13C or CO2 reference values are wrong. If set to true,
correct known incorrect values.}

\item{write_to_file}{Write calibrated ambient data to file?
(Mostly used for testing)}
}
\value{
Returns nothing to the environment, but creates a new output HDF5
        file containing calibrated carbon isotope values.
}
\description{
Use the gain-and-offset style calibration approach detailed in
Bowling et al. 2003 AFM. Wen et al. 2013 compared several different carbon
isotope calibration techniques and found this to be the superior method
under most circumstances. In brief, this method estimates gain and offset
parameters using linear regression on 12CO2 and 13CO2 isotopologues
separately. These gain and offset parameters are analogous to regression
slope and intercepts, but jointly correct for CO2 concentration dependence
and place d13C values on the VPDB scale. Gain and offset parameters are 
determined independently for each major isotopologue (12CO2 and 13CO2).
For the two reference materials
selected, the gain and offset parameters are defined by:
}
\details{
\deqn{G = (X_{2,ref}-X_{1,ref})/(X_{2,meas}-X_{1,meas})}
\deqn{O = X_{2,ref}- G X_{2,meas}}
Calibrated ambient isotopologues are then given as:
\deqn{X_{cal} = X_{meas} G + O}

Measurements of reference materials were considered "good" if the following
conditions were met:
\itemize{
  \item Measured CO2 concentrations were within 10 ppm
        of known "reference" concentrations.
  \item Variance of the CO2 concentration in standard peak was < 5 ppm.
  \item Measured d13C value must be within 5 per mil
        of known "reference" d13C value.
}
The first two criteria are intended to filter out periods where there is
a clear issue with the gas delivery system (i.e., nearly empty gas tank,
problem with a valve in the manifold, etc.); the third criterion was adopted
after visual inspection of data timeseries revealed that often the first
standard measurement following an instrument issue had higher-than-expected
error. This criterion clips clearly poor values.
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
