% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{calibrate_carbon_reference_data}
\alias{calibrate_carbon_reference_data}
\title{Title}
\usage{
calibrate_carbon_reference_data(inname, outname, standard, site, calDf)
}
\arguments{
\item{inname}{}

\item{outname}{}

\item{standard}{}

\item{site}{}

\item{calDf}{}
}
\value{

}
\description{
Title
}
