% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{calibrate_carbon_reference_data2}
\alias{calibrate_carbon_reference_data2}
\title{calibrate_carbon_reference_data2}
\usage{
calibrate_carbon_reference_data2(outname, standard, site, allData, calParams)
}
\arguments{
\item{outname}{Output file name.}

\item{standard}{Which standard are we working on? Must be "Low",
"Med", or "High"}

\item{site}{NEON 4-letter site code.}

\item{allData}{Uncalibrated reference data frames.}

\item{calParams}{Calibration data frame -
this is the output from fit_carbon_regression}
}
\value{
Nothing to the environment.
}
\description{
calibrate_carbon_reference_data2
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
\keyword{internal}
