% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_standards_carbon.R,
%   R/calibrate_standards_carbon_Bowling2003.R
\name{calibrate_standards_carbon}
\alias{calibrate_standards_carbon}
\title{Title}
\usage{
calibrate_standards_carbon(cal_df, ref_df, R_vpdb, f, r2_thres = 0.95)

calibrate_standards_carbon(cal_df, ref_df, R_vpdb, f, r2_thres = 0.95)
}
\arguments{
\item{cal_df}{}

\item{ref_df}{}

\item{R_vpdb}{}

\item{f}{}
}
\value{



}
\description{
Title

Title
}
