% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_standards_water.R
\name{calibrate_standards_water}
\alias{calibrate_standards_water}
\title{calibrate_standards_water}
\usage{
calibrate_standards_water(
  cal_df,
  ref_df,
  r2_thres = 0.95,
  correct_bad_refvals = FALSE,
  site,
  refGas
)
}
\arguments{
\item{cal_df}{Data.frame containing calibration parameters}

\item{ref_df}{Data.frame containing reference gas measurements}

\item{r2_thres}{Threshold for calibration regression to be used to
calibrate standards data. Default is 0.95. Calibrated reference
gas measurements occurring during calibration periods
with r2 values less than \code{r2_thres} will be marked NA.}

\item{correct_bad_refvals}{Should we correct known/suspected incorrect
reference values in the NEON HDF5 files? (Default = \code{FALSE}).}

\item{site}{Four letter NEON site code.
Only used if \code{correct_bad_refvals = TRUE}.}

\item{refGas}{One of "low", "med", or "high."
Only used if \code{correct_bad_refvals = TRUE}.}
}
\description{
calibrate_standards_water
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
