% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{calibrate_water_reference_data}
\alias{calibrate_water_reference_data}
\title{calibrate_water_reference_data}
\usage{
calibrate_water_reference_data(inname, outname, standard, site, calDf)
}
\arguments{
\item{inname}{Input file name.}

\item{outname}{Output file name.}

\item{standard}{Which standard are we working on? Must be "Low",
"Med", or "High"}

\item{site}{NEON 4-letter site code.}

\item{calDf}{Calibration data frame -
this is the output from fit_carbon_regression}
}
\value{
Nothing to the environment.
}
\description{
calibrate_water_reference_data
}
