% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{carbon_regression_plots}
\alias{carbon_regression_plots}
\title{Title}
\usage{
carbon_regression_plots(caldata, plot_filename, method)
}
\arguments{
\item{caldata}{}

\item{plot_filename}{}

\item{method}{}
}
\description{
Title
}
