% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_POSIXct_to_NEONhdf5_time.R
\name{convert_POSIXct_to_NEONhdf5_time}
\alias{convert_POSIXct_to_NEONhdf5_time}
\title{convert_POSIXct_to_NEONhdf5_time}
\usage{
convert_POSIXct_to_NEONhdf5_time(intime)
}
\arguments{
\item{intime}{}
}
\description{
convert_POSIXct_to_NEONhdf5_time
}
\examples{


}
