% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_qfqm_group.R
\name{copy_qfqm_group}
\alias{copy_qfqm_group}
\title{Title}
\usage{
copy_qfqm_group(data.list, outname, site, file, species)
}
\arguments{
\item{data.list}{}

\item{outname}{}

\item{site}{}

\item{file}{}

\item{species}{}
}
\value{

}
\description{
Title
}
