% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_ucrt_group.R
\name{copy_ucrt_group}
\alias{copy_ucrt_group}
\title{Title}
\usage{
copy_ucrt_group(data.list, outname, site, file, species)
}
\arguments{
\item{data.list}{}

\item{outname}{}

\item{site}{}

\item{file}{}
}
\description{
Title
}
