% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_corrections.R
\name{correct_carbon_ref_output}
\alias{correct_carbon_ref_output}
\title{Title}
\usage{
correct_carbon_ref_output(
  std_list,
  site,
  omit_already_corrected = TRUE,
  co2_tol = 5,
  d13c_tol = 0.25,
  refGas
)
}
\arguments{
\item{std_list}{}

\item{site}{}

\item{omit_already_corrected}{}

\item{co2_tol}{}

\item{d13c_tol}{}

\item{refGas}{}
}
\description{
Title
}
