% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_h5_groups_for_calibrated_isotopes.R
\name{create_h5_groups_for_calibrated_isotopes}
\alias{create_h5_groups_for_calibrated_isotopes}
\title{create_h5_groups_for_calibrated_isotopes}
\usage{
create_h5_groups_for_calibrated_isotopes(file, site)
}
\arguments{
\item{file}{Filename where new h5 groups should be created.}

\item{site}{Four-letter NEON site code for file.}
}
\value{
Nothing to the workspace, but adds new groups to
        output \code{file}.
}
\description{
create_h5_groups_for_calibrated_isotopes
}
\examples{

}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
