% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_NEON_data.R
\name{download_NEON_data}
\alias{download_NEON_data}
\title{Download NEON datasets from the API.}
\usage{
download_NEON_data(data.product, site.req = "all", site.month = "all")
}
\arguments{
\item{data.product}{NEON code of data product you'd like to retrieve.
For example, the eddy covariance bundle is DP4.00200.001.
(In this release, only DP4.00200.001 *should* work. Updates to fix soon.)}

\item{site.req}{Four-letter code of site you're requesting data from.
This can also be "all," which will download data from all
terrestrial sites.}

\item{site.month}{Month of data to retrieve, in yyyy-mm format. Can also
use "all," which will retrieve all available data for site.req.}
}
\value{
Requested data will be downloaded to the current working directory. 
A future release will allow more flexibility in download directory.
}
\description{
Download NEON datasets from the API.
}
\details{
Additional notes about different data products used here.

1) Eddy covariance data (DP4.00200.001) are stored on the NEON servers as 
monthly "basic" file or daily "extended" files. Currently, this script will
download the "basic" files and metadata files only. Extended files also include
footprint calculations, and can be added in a future release if needed (however,
this will require substantially more space to store). 

2) Meteorological variables are stored on NEON servers as separate CSV files based on:
a) position with respect to the tower (e.g., height on tower, or distance from tower) and
b) time resolution - I think usually this is 30 minute or 1 minute time resolution for most
products, but this has not been confirmed. Currently, for space/time considerations, 
this script will download *only* the 30 minute files. If 1 minute files are required later,
please submit a commit modifying this code or open an issue on the SPATIAL-Lab/NEONiso repository page.
}
\examples{
# get EC data from all sites from January 2018
download_NEON_data(data.product="DP4.00200.001",site.req="all",site.month="2018-01")
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
