% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_carbon_calibration_data.R,
%   R/extract_carbon_ucrt_data.R
\name{extract_carbon_calibration_data}
\alias{extract_carbon_calibration_data}
\title{Title}
\usage{
extract_carbon_calibration_data(ucrt.list, standard)

extract_carbon_calibration_data(ucrt.list, standard)
}
\arguments{
\item{standard}{}

\item{isoCo2.list}{}
}
\value{



}
\description{
Title

Title
}
