% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_data_extraction.R
\name{extract_water_calibration_data}
\alias{extract_water_calibration_data}
\title{extract_water_calibration_data}
\usage{
extract_water_calibration_data(data_list)
}
\arguments{
\item{data_list}{List containing data, from the /*/dp01/data/
group in NEON HDF5 file.}
}
\value{
Returns data frame of required variables.
}
\description{
extract_water_calibration_data
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
