% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_functions.R
\name{filter_median_Brock86}
\alias{filter_median_Brock86}
\title{Title}
\usage{
filter_median_Brock86(data, width = 7, threshold = 5)
}
\arguments{
\item{data}{}

\item{width}{}

\item{threshold}{}
}
\value{

}
\description{
Title
}
