% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_data_regression.R
\name{fit_carbon_regression}
\alias{fit_carbon_regression}
\title{Title}
\usage{
fit_carbon_regression(ref_data, method, calibration_half_width)
}
\arguments{
\item{ref_data}{}

\item{...}{}
}
\value{

}
\description{
Title
}
