% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_data_regression.R
\name{fit_water_regression}
\alias{fit_water_regression}
\title{fit_water_regression}
\usage{
fit_water_regression(stds, calibration_half_width, slope_tolerance, r2_thres)
}
\arguments{
\item{calibration_half_width}{}

\item{ref_data}{}

\item{method}{}
}
\value{

}
\description{
fit_water_regression
}
