% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_conversions.R
\name{get_Rstd}
\alias{get_Rstd}
\title{Title}
\usage{
get_Rstd(element)
}
\arguments{
\item{element}{}
}
\value{

}
\description{
Title
}
\examples{

}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
