% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_carbon_calibration_linreg.R
\name{process_carbon_calibration_data}
\alias{process_carbon_calibration_data}
\title{process_carbon_calibration_data}
\usage{
process_carbon_calibration_data(fname, site,
  time.diff.betweeen.standards = 1800, method = 1)
}
\arguments{
\item{fname}{}

\item{site}{}

\item{time.diff.betweeen.standards}{}

\item{method}{}
}
\description{
process_carbon_calibration_data
}
\examples{

}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
