% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_conversions.R
\name{R_to_delta}
\alias{R_to_delta}
\title{Title}
\usage{
R_to_delta(data_vector, element)
}
\arguments{
\item{data_vector}{}

\item{element}{}
}
\value{

}
\description{
Title
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
