% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_data.R
\name{restructure_ambient_data}
\alias{restructure_ambient_data}
\title{restructure_ambient_data}
\usage{
restructure_ambient_data(inpath, analyte)
}
\arguments{
\item{inpath}{Folder containing data to stack.}

\item{analyte}{Carbon (Co2) or water (H2o)?}
}
\value{
List of data extracted from files listed in inpath.
}
\description{
restructure_ambient_data
}
\keyword{internal}
