% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_standards_water.R
\name{restructure_reference_variables}
\alias{restructure_reference_variables}
\title{restructure_reference_variables}
\usage{
restructure_reference_variables(standard, varname)
}
\arguments{
\item{standard}{Which reference to apply function to? (high, med, low).
Needs to be a data.frame.}

\item{varname}{Which variable are we applying this function to? There's
a list of ~10 common ones to write to the hdf5 file.}
}
\value{
data.frame formatted for output to hdf5 file.
}
\description{
restructure_reference_variables
}
