% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_data.R
\name{restructure_variables}
\alias{restructure_variables}
\title{Restructures data.frames imported by ingest_data to shorten variable names
and  Wrapper function around restructure_carbon_variables
and restructure_water_variables.}
\usage{
restructure_variables(dataframe, varname, mode, group, species)
}
\arguments{
\item{dataframe}{Input data.frame, from \code{neonUtilities::stackEddy}}

\item{varname}{Which variable are we applying this function to? There's
a list of ~10 common ones to write to the hdf5 file.}

\item{mode}{Are we fixing a reference data frame or an ambient data frame?}

\item{group}{Data, ucrt, or qfqm?}

\item{species}{Set to 'Co2' for carbon; 'H2o' for water}
}
\value{
data.frame formatted for output to hdf5 file.
}
\description{
Restructures data.frames imported by ingest_data to shorten variable names
and  Wrapper function around restructure_carbon_variables
and restructure_water_variables.
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
