% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{setup_output_file}
\alias{setup_output_file}
\title{Title}
\usage{
setup_output_file(outname, site, analyte)
}
\arguments{
\item{outname}{}

\item{site}{}

\item{analyte}{}
}
\value{

}
\description{
Title
}
