% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_corrections.R
\name{swap_standard_isotoperatios}
\alias{swap_standard_isotoperatios}
\title{swap_standard_isotoperatios}
\usage{
swap_standard_isotoperatios(std_frame, dxs_thres = 500)
}
\arguments{
\item{std.frame}{Standard data frame to perform swap on.}

\item{dxs.thres}{d-excess threshold to indicate when to swap.}
}
\description{
There are a few suspected instances where the water
isotope ratios for oxygen and hydrogen have been flipped
in the reference data. This function corrects them until
they are corrected in the NEON database using a d-excess
filter.
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
