% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{terrestrial_core_sites}
\alias{terrestrial_core_sites}
\title{terrestrial_core_sites}
\usage{
terrestrial_core_sites()
}
\description{
Return a vector listing NEON core terrestrial sites.
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
