% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{terrestrial_gradient_sites}
\alias{terrestrial_gradient_sites}
\title{List terrestrial gradient sites}
\usage{
terrestrial_gradient_sites()
}
\value{
A vector listing NEON gradient terrestrial sites.
}
\description{
Returns a vector of four-letter NEON site codes for the
gradient terrestrial sites that have TIS instrumentation.
}
\examples{
terrestrial_gradient_sites()
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
