% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{validate_analyte}
\alias{validate_analyte}
\title{Standardize analyte names}
\usage{
validate_analyte(analyte)
}
\arguments{
\item{analyte}{Co2 or H2o?}
}
\value{
Standardized string for the water ('H2o') or
carbon ('Co2') systems to make sure strings
are standardized across package functions.
}
\description{
Enforces standard capitalization and formatting of H2o and Co2
analyte names across calibration functions.
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
