% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{validate_output_file}
\alias{validate_output_file}
\title{Validate output file.}
\usage{
validate_output_file(inname, outname, site, analyte)
}
\arguments{
\item{inname}{Input file name.}

\item{outname}{Output file name.}

\item{site}{NEON 4-letter site code.}

\item{analyte}{Carbon ('Co2') or water ('H2o') system?}
}
\value{
Nothing to environment, simply checks to make sure expected groups
are in output.
}
\description{
Function ensures that the output file has the correct
groups in it, as a check to ensure proper file structure at the
end of the calibration routines.
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
