% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_diagnostic_package.R
\name{water_diagnostic_package_bymonth}
\alias{water_diagnostic_package_bymonth}
\title{water_diagnostic_package_bymonth.R}
\usage{
water_diagnostic_package_bymonth(data_path, plot_path, which_sites = "all")
}
\arguments{
\item{data_path}{Provide path to where calibrated data from one site live.}

\item{plot_path}{Path to where output pdf plots should be written.}

\item{which_sites}{Which NEON sites to run plots for? Default = all}
}
\description{
water_diagnostic_package_bymonth.R
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
