% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_diagnostic_package.R
\name{water_diagnostic_package_bysite}
\alias{water_diagnostic_package_bysite}
\title{water_diagnostic_package_bysite.R}
\usage{
water_diagnostic_package_bysite(data_path, plot_path, which_sites = "all")
}
\arguments{
\item{data_path}{Provide path to where calibrated data live.
Assumes that the \code{calibrate_water_linreg_bysite} function has
been used to calibrate these data, and there is one file per site.}

\item{plot_path}{Path to where output pdf plots should be written.}

\item{which_sites}{Which NEON sites to run plots for? Default = all}
}
\description{
water_diagnostic_package_bysite.R
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
