% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{water_isotope_sites}
\alias{water_isotope_sites}
\title{water_isotope_sites}
\usage{
water_isotope_sites()
}
\description{
Return a vector listing NEON sites measuring water vapor isotope ratios.
}
\examples{
water_isotope_sites()
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
