% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{write_carbon_ambient_data}
\alias{write_carbon_ambient_data}
\title{Title}
\usage{
write_carbon_ambient_data(outname, site, amb_data_list)
}
\arguments{
\item{outname}{}

\item{site}{}

\item{amb_data_list}{}
}
\value{

}
\description{
Title
}
