% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{write_carbon_calibration_data}
\alias{write_carbon_calibration_data}
\title{Title}
\usage{
write_carbon_calibration_data(outname, site, calDf, method)
}
\arguments{
\item{outname}{}

\item{calDf}{}

\item{method}{}
}
\value{

}
\description{
Title
}
