% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{write_carbon_reference_data2}
\alias{write_carbon_reference_data2}
\title{write_carbon_reference_data2}
\usage{
write_carbon_reference_data2(outname, site, allData, calDf)
}
\arguments{
\item{outname}{Output file name.}

\item{site}{NEON 4-letter site code.}

\item{allData}{Uncalibrated reference data frames.}

\item{calDf}{Calibration data frame -
this is the output from fit_carbon_regression}
}
\value{
Nothing to the environment, but writes calibrated reference data to hdf5 file.
}
\description{
write_carbon_reference_data2
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
\keyword{internal}
