% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{write_ucrt}
\alias{write_ucrt}
\title{Title}
\usage{
write_ucrt(inname, outname, site, analyte)
}
\arguments{
\item{inname}{}

\item{outname}{}

\item{site}{}

\item{analyte}{}
}
\value{

}
\description{
Title
}
