% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{write_water_ambient_data}
\alias{write_water_ambient_data}
\title{Write calibrated ambient water isotope ratio observations to file.}
\usage{
write_water_ambient_data(outname, site, amb_data_list, fid = NULL)
}
\arguments{
\item{outname}{Output file name.}

\item{site}{NEON 4-letter site code.}

\item{amb_data_list}{Calibrated list of ambient data -
this is the output from one of the calibrate_ambient_water* functions.}

\item{fid}{Optional open HDF5 file handle. If NULL, the file is
opened and closed internally.}
}
\value{
Nothing to the environment, but writes data in amb_data_list to file.
}
\description{
Write out ambient observations from the NEON EC
towers where the isotope data
have been calibrated using this package.
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
