% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{write_water_calibration_data}
\alias{write_water_calibration_data}
\title{Write water calibration parameters to file}
\usage{
write_water_calibration_data(outname, site, cal_df, fid = NULL)
}
\arguments{
\item{outname}{Output file name.}

\item{site}{NEON 4-letter site code.}

\item{cal_df}{Calibration data frame -
this is the output from fit_water_regression}

\item{fid}{Optional open HDF5 file handle. If NULL, the file is
opened and closed internally.}
}
\value{
Nothing to the environment, but writes out the
calibration parameters (e.g.,
regression slopes and intercepts) to the output
hdf5 file.
}
\description{
Write a \code{data.frame} with slope, intercepts, and error estimates of
calibrations for water isotope system.
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
