% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{write_water_reference_data}
\alias{write_water_reference_data}
\title{write_water_reference_data}
\usage{
write_water_reference_data(inname, outname, site, lowDf, medDf, highDf, calDf)
}
\arguments{
\item{inname}{Input file name.}

\item{outname}{Output file name.}

\item{site}{NEON 4-letter site code.}

\item{lowDf}{Dataframe corresponding to the "low" reference water.}

\item{medDf}{Data frame corresponding to the "med" reference water.}

\item{highDf}{Data frame corresponding to the "high" reference water.}

\item{calDf}{Calibration data frame -
this is the output from fit_water_regression}
}
\value{
Nothing to the environment, but writes calibrated
reference data to hdf5 file.
}
\description{
write_water_reference_data
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
