% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_ambient_carbon_Bowling2003.R
\name{calibrate_ambient_carbon_Bowling2003}
\alias{calibrate_ambient_carbon_Bowling2003}
\title{calibrate_ambient_carbon_Bowling2003}
\usage{
calibrate_ambient_carbon_Bowling2003(amb.data.list, caldf, outname, site,
  file, method, forceToEnd = TRUE, forceToBeginning = TRUE)
}
\arguments{
\item{amb.data.list}{}

\item{caldf}{}

\item{outname}{}

\item{site}{}

\item{file}{}

\item{method}{}

\item{force.to.end}{}

\item{force.to.beginning}{}
}
\description{
calibrate_ambient_carbon_Bowling2003
}
\examples{

}
