% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_ambient_carbon_Bowling2003.R
\name{calibrate_ambient_carbon_Bowling2003}
\alias{calibrate_ambient_carbon_Bowling2003}
\title{calibrate_ambient_carbon_Bowling2003}
\usage{
calibrate_ambient_carbon_Bowling2003(
  amb.data.list,
  caldf,
  outname,
  site,
  file,
  filter.data
)
}
\arguments{
\item{amb.data.list}{}

\item{caldf}{Calibration data frame containing gain and offset values for 12C and 13C isotopologues.}

\item{outname}{Output file name.}

\item{site}{Four-letter NEON code corersponding to site being processed.}

\item{file}{}

\item{filter.data}{}
}
\value{
Nothing to environment; returns calibrated ambient observations to the calibrate_carbon_Bowling2003 function. This function is not designed to be called on its own.
}
\description{
calibrate_ambient_carbon_Bowling2003
}
\examples{

}
