% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_ambient_carbon_linreg.R
\name{calibrate_ambient_carbon_linreg}
\alias{calibrate_ambient_carbon_linreg}
\title{calibrate_ambient_carbon_isotopes}
\usage{
calibrate_ambient_carbon_linreg(
  amb_data_list,
  caldf,
  outname,
  site,
  file,
  force_to_end = TRUE,
  force_to_beginning = TRUE,
  r2_thres
)
}
\arguments{
\item{caldf}{Calibration data frame containing gain and offset values for
12C and 13C isotopologues.}

\item{outname}{Output variable name. Inherited from
\code{calibrate_ambient_carbon_linreg}}

\item{site}{Four-letter NEON code corersponding to site being processed.}

\item{file}{Output file name. Inherited from
\code{calibrate_ambient_carbon_linreg}}

\item{amb.data.list}{List containing an ambient d13C dataset.
Will include all variables in 000_0x0_xxm. (character)}

\item{force.to.end}{In given month, calibrate ambient data later than last
calibration, using the last calibration? (default true)}

\item{force.to.beginning}{In given month, calibrate ambient data before than
first calibration, using the first calibration? (default true)}

\item{r2.thres}{Minimum r2 value for calibration to be considered "good" and
applied to ambient data.}
}
\value{
Nothing to environment; returns calibrated ambient observations to
    the output file. This function is not designed to be called on its own.
}
\description{
calibrate_ambient_carbon_isotopes
}
\examples{

}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}

Function called by \code{calibrate_ambient_carbon_linreg} to apply
gain and offset parameters to the ambient datasets (000_0x0_09m and
000_0x0_30m). This function should generally not be used independently,
but should be used with \code{calibrate_ambient_carbon_linreg}.
}
