% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_ambient_water_linreg.R
\name{calibrate_ambient_water_linreg}
\alias{calibrate_ambient_water_linreg}
\title{calibrate_ambient_water_isotopes}
\usage{
calibrate_ambient_water_linreg(
  amb_data_list,
  caldf,
  outname,
  site,
  file,
  force_to_end = TRUE,
  force_to_beginning = TRUE
)
}
\arguments{
\item{force.to.beginning}{}
}
\value{

}
\description{
calibrate_ambient_water_isotopes
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
