% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_carbon_Bowling2003.R
\name{calibrate_carbon_Bowling2003}
\alias{calibrate_carbon_Bowling2003}
\title{calibrate_carbon_Bowling2003}
\usage{
calibrate_carbon_Bowling2003(
  inname,
  outname,
  site,
  time.diff.between.standards = 1800,
  force.cal.to.beginning = TRUE,
  force.cal.to.end = TRUE,
  ucrt.source = "NEON"
)
}
\arguments{
\item{inname}{Name of the input file.}

\item{outname}{Name of the output file.}

\item{site}{Four letter NEON site code for site being processed.}

\item{time.diff.between.standards}{Time (in seconds) required between consecutive standard measurements.}

\item{force.cal.to.beginning}{Extend first calibration to the beginning of the file? (Default true)}

\item{force.cal.to.end}{Extend last calibration to the end of the file? (Detault true)}

\item{ucrt.source}{Where do we take uncertainty estimates from? (not used currently, set to NEON)}
}
\value{
Returns nothing to the workspace, but creates a new output file containing calibrated carbon isotope values.
}
\description{
Use the gain-and-offset style calibration approach detailed in Bowling et al. 2003 AFM.
Wen et al. 2011 compared several different carbon isotope calibration techniques and
found this to be the superior method under most circumstances.
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
