% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_carbon_linreg.R
\name{calibrate_carbon_linreg}
\alias{calibrate_carbon_linreg}
\title{calibrate_carbon_linreg}
\usage{
calibrate_carbon_linreg(
  inname,
  outname,
  site,
  time_diff_between_standards = 1800,
  force_cal_to_beginning = TRUE,
  force_cal_to_end = TRUE,
  r2_thres = 0.95
)
}
\arguments{
\item{inname}{Name of the input file. (character)}

\item{outname}{Name of the output file. (character)}

\item{site}{Four-letter NEON code for site being processed. (character)}

\item{force.cal.to.beginning}{Extend first calibration to
the beginning of the file?}

\item{time.diff.between.standards}{Time (in seconds) required between
consecutive standard measurements.
Used to define a calibration "period."}

\item{force.cal.to.end}{Extend last calibration to the end of the file?}
}
\value{
nothing to the workspace, but creates a new output file of
        calibrated carbon isotope data.
}
\description{

}
\details{

}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
