% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_carbon_linreg.R
\name{calibrate_carbon_linreg}
\alias{calibrate_carbon_linreg}
\title{calibrate_carbon_linreg}
\usage{
calibrate_carbon_linreg(
  inname,
  outname,
  site,
  time_diff_between_standards = 1800,
  force_cal_to_beginning = TRUE,
  force_cal_to_end = TRUE,
  r2_thres = 0.95,
  correct_refData = TRUE,
  write_to_file = TRUE
)
}
\arguments{
\item{inname}{Name of the input file. (character)}

\item{outname}{Name of the output file. (character)}

\item{site}{Four-letter NEON code for site being processed. (character)}

\item{time_diff_between_standards}{Time (in seconds) required between
consecutive standard measurements.
Used to define a calibration "period."}

\item{force_cal_to_beginning}{Extend first calibration to
the beginning of the file?}

\item{force_cal_to_end}{Extend last calibration to the end of the file?}

\item{r2_thres}{Minimum r2 threshold of an "acceptable" calibration. Acts to
remove calibration periods where a measurement error makes
relationship nonlinear. Default = 0.95}

\item{correct_refData}{Should we replace known/suspected incorrect reference
values in the NEON HDF5 files? If \code{TRUE} (default), then
corrects values using a function in \code{standard_corrections.R}.}

\item{write_to_file}{Write calibrated ambient data to file?
(Mostly used for testing)}
}
\value{
nothing to the workspace, but creates a new output file of
        calibrated carbon isotope data.
}
\description{
This function will calibrate NEON carbon isotope data using an ordinary
least squares linear regression between measured and reference d13C values.
Correction equation is determined by regressing the reference values on
the measured values, and uses this equation to calibrate the ambient data.
In brief, this function takes the following steps:
}
\details{
\enumerate{
  \item Extracts calibration data from uncalibrated file.
  \item Basic QA/QC on each calibration data point, where the following
        factors must be true:
  \itemize{
    \item Calibration "peak" must have >= 200 data points, to remove some
          observed issues with gas manifold valves.
    \item Calibration "peak" must not be missing.
    \item Only one value per day meeting these criteria are selected.
  }
  \item Calibration periods are defined to bracket a sample of ambient data.
        In many cases, this will be one day bracketed by measurements of
       reference materials immediately before after this period.
  \item Determine the slope, intercept, and r^2 of a regression of each
        calibration period. Calibration error is estimated using
        the difference between the model-predicted reference value of each
        standard compared to the "known" reference value
        for the "medium" standard only. As this bracketing calibration may
        use two measurements of the medium standard, only the maximum
        difference value is retained as the more conservative approach.

  \item Regression parameters are written to a dataset in a new output file
        under \code{/site/dp01/data/isoCo2/calData/calRegressions}
  \item Regression equations are applied ambient data,
        and written to same new output file.
}

The qfqm and ucrt folders are also copied over from the original file,
and are unchanged.
}
\examples{
fin <- system.file("NEON_sample_packed.h5", package = "NEONiso")
# note typical use case would have write_to_file = TRUE
calibrate_carbon_linreg(inname = fin, outname = "example.h5", 
                            site = "ONAQ", write_to_file = FALSE)
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
