% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_standards_carbon.R
\name{calibrate_standards_carbon}
\alias{calibrate_standards_carbon}
\title{calibrate_standards_carbon}
\usage{
calibrate_standards_carbon(cal_df, ref_df, R_vpdb, f, r2_thres = 0.95)
}
\arguments{
\item{cal_df}{Data.frame containing calibration parameters}

\item{ref_df}{Data.frame containing reference gas measurements}

\item{R_vpdb}{Ratio of 13C to 12C in Vienna Pee Dee Belemnite, inhereted
from the script calling this function.}

\item{f}{Fraction of CO2 isotopologues that are not 12CO2 or 13CO2. 
Inhereted from script calling this function.}
}
\value{

}
\description{
calibrate_standards_carbon
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
