% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_standards_water.R
\name{calibrate_standards_water}
\alias{calibrate_standards_water}
\title{calibrate_standards_water}
\usage{
calibrate_standards_water(cal_df, ref_df, r2_thres = 0.95)
}
\arguments{
\item{cal_df}{Data.frame containing calibration parameters}

\item{ref_df}{Data.frame containing reference gas measurements}

\item{r2_thres}{Threshold for calibration regression to be used to 
calibrate standards data. Default is 0.95. Calibrated reference
gas measurements occuring during calibration periods
with r2 values less than \code{r2_thres} will be marked NA.}

\item{R_vpdb}{Ratio of 13C to 12C in Vienna Pee Dee Belemnite, inhereted
from the script calling this function.}

\item{f}{Fraction of CO2 isotopologues that are not 12CO2 or 13CO2. 
Inhereted from script calling this function.}
}
\value{

}
\description{
calibrate_standards_water
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
