% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_water_linreg.R
\name{calibrate_water_linreg}
\alias{calibrate_water_linreg}
\title{calibrate_water_linreg}
\usage{
calibrate_water_linreg(
  inname,
  outname,
  site,
  time_diff_betweeen_standards = 1800
)
}
\arguments{
\item{inname}{Name of the input file.}

\item{outname}{Name of the output file.}

\item{site}{Four-letter NEON code for site being processed.}

\item{time_diff_betweeen_standards}{Time (in seconds) required between consecutive standard measurements.}
}
\value{
nothing to the workspace, but creates a new output file of
        calibrated carbon isotope data.
}
\description{
calibrate_water_linreg
}
\examples{

}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
