% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_water_linreg.R
\name{calibrate_water_linreg}
\alias{calibrate_water_linreg}
\title{calibrate_water_linreg}
\usage{
calibrate_water_linreg(
  inname,
  outname,
  site,
  time.diff.betweeen.standards = 1800,
  force.cal.to.beginning = TRUE,
  force.cal.to.end = TRUE
)
}
\arguments{
\item{inname}{Name of the input file.}

\item{outname}{Name of the output file.}

\item{site}{Four-letter NEON code for site being processed.}

\item{time.diff.betweeen.standards}{Time (in seconds) required between consecutive standard measurements.}

\item{force.cal.to.beginning}{Extend first calibration to the beginning of the file? (Default true)}

\item{force.cal.to.end}{Extend last calibration to the end of the file? (Detault true)}
}
\value{

}
\description{
calibrate_water_linreg
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
