% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_diagnostic_package.R
\name{carbon_diagnostic_package}
\alias{carbon_diagnostic_package}
\title{Title}
\usage{
carbon_diagnostic_package(data_path, plot_path, which.sites = "all")
}
\arguments{
\item{data_path}{Provide path to where calibrated data from one site live.}

\item{plot_path}{}

\item{method}{Bowling or linear regression calibration method? (Valid values: Bowling, LinReg)}
}
\value{

}
\description{
Title
}
\examples{

}
