% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{carbon_regression_plots}
\alias{carbon_regression_plots}
\title{Make plots of carbon calibration data for debugging}
\usage{
carbon_regression_plots(caldata, plot_filename, method, mtitle)
}
\arguments{
\item{caldata}{Data frame corresponding to a specific calibration
period.}

\item{plot_filename}{What should the output file name for diagnostic plot be?}

\item{method}{Which method are we using? Currently works for gain/offset.}

\item{mtitle}{Fed from above routine - what should the plot title be?}
}
\value{
Nothing to the environment, but a pdf plot to a file.
}
\description{
Makes plots of carbon calibration data regressions, primarily for
debugging and validation purposes.
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
