% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_functions.R
\name{convert_NEONhdf5_to_POSIXct_time}
\alias{convert_NEONhdf5_to_POSIXct_time}
\title{convert_NEONhdf5_to_POSIXct_time}
\usage{
convert_NEONhdf5_to_POSIXct_time(intime)
}
\arguments{
\item{intime}{Vector of datetimes in NEON data files (as string)
to convert to POSIXct class}
}
\value{
Vector of datetimes from NEON data file now in POSIXct format.
}
\description{
convert_NEONhdf5_to_POSIXct_time
}
\examples{
convert_NEONhdf5_to_POSIXct_time("2019-06-01T12:00:00.000Z")
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
