% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_POSIXct_to_NEONhdf5_time.R
\name{convert_POSIXct_to_NEONhdf5_time}
\alias{convert_POSIXct_to_NEONhdf5_time}
\title{convert_POSIXct_to_NEONhdf5_time}
\usage{
convert_POSIXct_to_NEONhdf5_time(intime)
}
\arguments{
\item{intime}{Time to convert from NEON HDF5 files. Should be a character string.}
}
\value{
Returns POSIXct version of NEON time variable.
}
\description{
convert_POSIXct_to_NEONhdf5_time
}
\examples{


}
