% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{copy_qfqm_group}
\alias{copy_qfqm_group}
\title{copy_qfqm_group}
\usage{
copy_qfqm_group(data_list, outname, site, file, species)
}
\arguments{
\item{data_list}{List of groups to retrieve qfqm data from.}

\item{outname}{Output filename.}

\item{site}{Four-letter NEON site code.}

\item{file}{Input filename.}

\item{species}{CO2 or H2O? Same function used for both CO2 and H2O isotopes.}
}
\value{
Nothing to the workspace, but copies qfqm group from input file to
output file.
}
\description{
copy_qfqm_group
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
