% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{copy_ucrt_group}
\alias{copy_ucrt_group}
\title{copy_ucrt_group}
\usage{
copy_ucrt_group(data_list, outname, site, file, species)
}
\arguments{
\item{data_list}{List of groups to retrieve ucrt data from.}

\item{outname}{Output file name.}

\item{site}{NEON 4-letter site code.}

\item{file}{Input file name.}

\item{species}{H2O or CO2.}
}
\value{
Nothing to the workspace, but copies ucrt group from input file to
output file.
}
\description{
copy_ucrt_group
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
