% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_h5_groups_for_calibrated_isotopes.R
\name{create_h5_groups_for_calibrated_isotopes}
\alias{create_h5_groups_for_calibrated_isotopes}
\title{create_h5_groups_for_calibrated_isotopes}
\usage{
create_h5_groups_for_calibrated_isotopes(file, site)
}
\arguments{
\item{file}{}

\item{site}{}
}
\value{

}
\description{
create_h5_groups_for_calibrated_isotopes
}
\examples{


}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
