% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_conversions.R
\name{delta_to_R}
\alias{delta_to_R}
\title{Title}
\usage{
delta_to_R(data_vector, element)
}
\arguments{
\item{data_vector}{}

\item{element}{}
}
\value{

}
\description{
Title
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
