% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_NEON_data.R
\name{download_NEON_data}
\alias{download_NEON_data}
\title{Download NEON datasets from the API.}
\usage{
download_NEON_data(data.product = "DP4.00200.001", site.req = "all",
  site.month = "all")
}
\arguments{
\item{data.product}{NEON code of data product you'd like to retrieve}

\item{site.req}{Four-letter code of site that you'd like to retrieve data from.}

\item{site.month}{Which site month would you like to download?}
}
\description{
Download NEON datasets from the API.
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
