% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_data_regression.R
\name{estimate_calibration_error}
\alias{estimate_calibration_error}
\title{estimate_calibration_error}
\usage{
estimate_calibration_error(formula, data)
}
\arguments{
\item{formula}{Formula to pass to caret::train to perform cross validation.}

\item{data}{Data frame to perform cross-validation on.}
}
\description{
estimate_calibration_error
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
