% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_carbon_calibration_data.R
\name{extract_carbon_calibration_data}
\alias{extract_carbon_calibration_data}
\title{Title}
\usage{
extract_carbon_calibration_data(isoCo2.list, standard)
}
\arguments{
\item{isoCo2.list}{}

\item{standard}{}
}
\description{
Title
}
