% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_data_extraction.R
\name{extract_carbon_calibration_data}
\alias{extract_carbon_calibration_data}
\title{extract_carbon_calibration_data.R}
\usage{
extract_carbon_calibration_data(data_list)
}
\arguments{
\item{data_list}{List containing data, from the /*/dp01/data/
group in NEON HDF5 file.}

\item{standard}{String indicating whether to grab data from the high,
medium, or low standard.}

\item{input_stackEddy}{Was the data provided in data_list argument
generated by neonUtilities::stackEddy (set to TRUE) 
or rhdf5::h5read (set to FALSE)? Default FALSE.}
}
\value{
Returns data frame of required variables.
}
\description{
extract_carbon_calibration_data.R
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
