% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_carbon_calibration_data.R
\name{extract_carbon_calibration_data}
\alias{extract_carbon_calibration_data}
\title{extract_carbon_calibration_data.R}
\usage{
extract_carbon_calibration_data(
  data_list,
  ucrt_list,
  standard,
  ucrt_source = "data"
)
}
\arguments{
\item{standard}{String indicating whether to grab data from the high,
medium, or low standard.}

\item{data.list}{List containing data, from the /*/dp01/data/
group in NEON HDF5 file.}

\item{ucrt.list}{List containing uncertainty data, from the /*/dp01/ucrt/
group in NEON HDF5 file.}

\item{ucrt.source}{Where from HDF5 file should  variance be extracted from?
(Valid values are "data" and "ucrt")}
}
\value{
Returns data frame of required variables.
}
\description{
extract_carbon_calibration_data.R
}
