% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_functions.R
\name{filter_median_brock86}
\alias{filter_median_brock86}
\title{Apply a median absolute deviation filter}
\usage{
filter_median_brock86(data, width = 7, threshold = 5)
}
\arguments{
\item{data}{Vector to filter.}

\item{width}{Width of filter, in rows.}

\item{threshold}{Only filter values that are \code{abs(threshold)}
away from median}
}
\value{
Returns filtered vector.
}
\description{
Median absolute deviation filter of Brock 1986, with user specified
width and magnitude thresholds.
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
