% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_functions.R
\name{filter_median_Brock86}
\alias{filter_median_Brock86}
\title{filter_median_Brock86}
\usage{
filter_median_Brock86(data, width = 7, threshold = 5)
}
\arguments{
\item{data}{Vector to filter.}

\item{width}{Width of filter, in rows.}

\item{threshold}{Only filter values that are \code{abs(threshold)}
away from median}
}
\value{
Returns filtered vector.
}
\description{
Median absolute deviation filter of Brock 1986.
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
