% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_conversions.R
\name{get_Rstd}
\alias{get_Rstd}
\title{Return heavy-to-light isotope ratio of primary standard.}
\usage{
get_Rstd(element)
}
\arguments{
\item{element}{Which element to return standard ratio -
carbon, oxygen, or hydrogen.}
}
\value{
Heavy-to-light isotope ratio of most common
stable isotope standard. VSMOW for water,
VPDB for carbon.
}
\description{
Returns the heavy-to-light isotope ratio of the dominant
standard for that element. Vienna Standard Mean Ocean Water
(VSMOW) for oxygen and hydrogen isotopes, Vienna Pee Dee
Belemnite (VPDB) for carbon stable isotopes.
}
\examples{
get_Rstd("carbon") # returns 0.0111797
get_Rstd("oxygen") # returns 2005.20e-6

}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
