% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_conversions.R
\name{get_Rstd}
\alias{get_Rstd}
\title{get_Rstd}
\usage{
get_Rstd(element)
}
\arguments{
\item{element}{Which element to return standard ratio - 
carbon, oxygen, or hydrogen.}
}
\value{
Heavy-to-light isotope ratio of most common
        stable isotope standard. VMOW for water,
        VPDB for carbon.
}
\description{
get_Rstd
}
\examples{
get_Rstd('carbon') # returns 0.0111797
get_Rstd('oxygen') # returns 2005.2e-6
get_Rstd('hydrogen') # returns 155.76e-6

}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
