% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_data.R
\name{ingest_data}
\alias{ingest_data}
\title{ingest_data}
\usage{
ingest_data(inname, analyte)
}
\arguments{
\item{inname}{A file (or list of files) to extract data from for calibration.}

\item{analyte}{Carbon (Co2) or water (H2o)?}
}
\value{
List of data frames, taken from files specified in \code{inname}
}
\description{
ingest_data
}
