% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_data_regression.R
\name{loocv}
\alias{loocv}
\title{Leave-one-out cross validation}
\usage{
loocv(mod)
}
\arguments{
\item{mod}{Fitted model to estimate leave-one-out CV on.}
}
\description{
Calculate analytic leave-one-out cross variance error estimate
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
