% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{manage_local_EC_archive}
\alias{manage_local_EC_archive}
\title{Title}
\usage{
manage_local_EC_archive(file_dir, get = TRUE, trim = FALSE)
}
\arguments{
\item{file_dir}{}

\item{get}{}

\item{trim}{}
}
\value{

}
\description{
Title
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
