% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEONiso-package.R
\docType{package}
\name{NEONiso-package}
\alias{NEONiso}
\alias{NEONiso-package}
\title{NEONiso}
\description{
Functions for downloading,
calibrating, and analyzing atmospheric isotope data bundled
into the eddy covariance data products of the National Ecological
Observatory Network (NEON)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lanl/NEONiso}
  \item \url{https://lanl.github.io/NEONiso/}
  \item Report bugs at \url{https://github.com/lanl/NEONiso/issues}
}

}
\author{
\strong{Maintainer}: Rich Fiorella \email{rfiorella@lanl.gov} (\href{https://orcid.org/0000-0002-0824-4777}{ORCID})

}
\keyword{internal}
