% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_conversions.R
\name{R_to_delta}
\alias{R_to_delta}
\title{R_to_delta}
\usage{
R_to_delta(data_vector, element)
}
\arguments{
\item{data_vector}{A vector of isotope ratios (e.g., R values).}

\item{element}{Which element to return delta values - 
carbon, oxygen, or hydrogen.}
}
\value{
Vector of isotope ratios in delta notation.
}
\description{
R_to_delta
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
