% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{setup_output_file}
\alias{setup_output_file}
\title{Structure a new HDF5 file}
\usage{
setup_output_file(inname, outname, site, analyte)
}
\arguments{
\item{inname}{Input file name.}

\item{outname}{Output file name.}

\item{site}{NEON 4-letter site code.}

\item{analyte}{Carbon ('Co2') or water ('H2o') system?}
}
\value{
Nothing to the environment, but creates a new data file
with the most basic output HDF5 structure consistent with
NEON's data files.
}
\description{
Creates a skeleton HDF5 file for the calibrated data,
essentially setting up the HDF5 groups at the /site/dp01/\{data,ucrt,qfqm\}
level.
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
