% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{terrestrial_core_sites}
\alias{terrestrial_core_sites}
\title{List terrestrial core sites}
\usage{
terrestrial_core_sites()
}
\value{
A vector listing NEON core terrestrial sites.
}
\description{
Returns a vector of four-letter NEON site codes for the
core terrestrial sites that have TIS instrumentation.
}
\examples{
terrestrial_core_sites()
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
