% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrestrial_core_sites.R
\name{terrestrial_core_sites}
\alias{terrestrial_core_sites}
\title{terrestrial_core_sites}
\usage{
terrestrial_core_sites()
}
\description{
terrestrial_core_sites
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
