% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{terrestrial_core_sites}
\alias{terrestrial_core_sites}
\title{terrestrial_core_sites}
\usage{
terrestrial_core_sites()
}
\value{
A vector listing NEON core terrestrial sites.
}
\description{
terrestrial_core_sites
}
\examples{
terrestrial_core_sites()
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
