% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{validate_analyte}
\alias{validate_analyte}
\title{validate_analyte}
\usage{
validate_analyte(analyte)
}
\arguments{
\item{analyte}{Co2 or H2o?}
}
\value{
Standardized string for the water ('H2o') or
        carbon ('Co2') systems to make sure strings
        are standardized across package functions.
}
\description{
validate_analyte
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
