% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{water_isotope_sites}
\alias{water_isotope_sites}
\title{List sites with water vapor isotope ratios.}
\usage{
water_isotope_sites()
}
\value{
A vector listing NEON sites measuring water vapor isotope ratios.
}
\description{
Returns a vector of four-letter NEON site codes for the
terrestrial sites that have water vapor isotope ratio instrumentation.
}
\examples{
water_isotope_sites()
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
