% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{write_carbon_ambient_data}
\alias{write_carbon_ambient_data}
\title{Write calibrated carbon ambient data to file}
\usage{
write_carbon_ambient_data(
  outname,
  site,
  amb_data_list,
  to_file = TRUE,
  fid = NULL
)
}
\arguments{
\item{outname}{Output file name.}

\item{site}{NEON 4-letter site code.}

\item{amb_data_list}{Calibrated list of ambient data -
this is the output from one of the calibrate_ambient_carbon* functions.}

\item{to_file}{Write to file (TRUE) or to environment (FALSE).}

\item{fid}{Optional open HDF5 file handle. If NULL, the file is
opened and closed internally.}
}
\value{
Nothing to the environment, but writes data in amb_data_list to file.
}
\description{
Write out ambient observations from the NEON EC
towers where the isotope data (either H2O or CO2)
have been calibrated using this package.
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
