% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{write_carbon_calibration_data}
\alias{write_carbon_calibration_data}
\title{Write carbon calibrations to file}
\usage{
write_carbon_calibration_data(
  outname,
  site,
  cal_df,
  method,
  to_file = TRUE,
  fid = NULL
)
}
\arguments{
\item{outname}{Output file name.}

\item{site}{NEON 4-letter site code.}

\item{cal_df}{Calibration data frame -
this is the output from fit_carbon_regression}

\item{method}{Was the Bowling et al. 2003 or the linear regression
method used in fit_carbon_regression?}

\item{to_file}{Write to file (TRUE) or to environment (FALSE).}

\item{fid}{Optional open HDF5 file handle. If NULL, the file is
opened and closed internally.}
}
\value{
Nothing to the environment, but writes out the
calibration parameters (e.g., gain and offset or
regression slopes and intercepts) to the output
hdf5 file.
}
\description{
Write a \code{data.frame} with slope, intercepts, and error estimates of
calibrations for carbon isotope system. If \code{gainoffset} method was used
the slopes/intercepts are called gain/offsets for each isotopologue.
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
