% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{write_qfqm}
\alias{write_qfqm}
\title{write_qfqm}
\usage{
write_qfqm(inname, outname, site, analyte)
}
\arguments{
\item{inname}{Input file name.}

\item{outname}{Output file name.}

\item{site}{NEON 4-letter site code.}

\item{analyte}{Carbon ('Co2') or water ('H2o') system?}
}
\value{
Nothing to the environment, but writes qfqm data to file.
}
\description{
Write NEON's qfqm data for an isotope species to
output file. Wraps copy_qfqm_group.
}
\author{
Rich Fiorella \email{rich.fiorella@utah.edu}
}
