% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excised.R
\name{write_ucrt}
\alias{write_ucrt}
\title{write_ucrt}
\usage{
write_ucrt(inname, outname, site, analyte)
}
\arguments{
\item{inname}{Input file name.}

\item{outname}{Output file name.}

\item{site}{NEON 4-letter site code.}

\item{analyte}{Carbon ('Co2') or water ('H2o') system?}
}
\value{
Nothing to the environment, but writes ucrt data to file.
}
\description{
Write NEON's ucrt data for an isotope species to
output file. Wraps copy_ucrt_group.
}
\author{
Rich Fiorella \email{rfiorella@lanl.gov}
}
